+++ Etyllica/src/br/com/etyllica/core/event/PointerEvent.java
@@ -16,9 +16,9 @@ public class PointerEvent {
 	
 	private PointerState state;
 	
-	private float x;
+	private int x;
 	
-	private float y;
+	private int y;
 	
 	private int amount;
 	
@@ -67,19 +67,19 @@ public class PointerEvent {
 		this.state = state;
 	}
 
-	public float getX() {
+	public int getX() {
 		return x;
 	}
 
-	public void setX(float x) {
+	public void setX(int x) {
 		this.x = x;
 	}
 
-	public float getY() {
+	public int getY() {
 		return y;
 	}
 
-	public void setY(float y) {
+	public void setY(int y) {
 		this.y = y;
 	}
