+++ Etyllica/src/br/com/etyllica/core/video/Monitor.java
@@ -2,7 +2,7 @@ package br.com.etyllica.core.video;
 
 public class Monitor extends br.com.etyllica.layer.GeometricLayer{
 
-	public Monitor(float x, float y, float w, float h){
+	public Monitor(int x, int y, int w, int h){
 		super(x,y,w,h);
 	}
