+++ Etyllica/src/br/com/etyllica/gui/icon/DownArrow.java
@@ -2,15 +2,14 @@ package br.com.etyllica.gui.icon;
 
 public class DownArrow extends PolygonalIcon{
 
-	public DownArrow(float x, float y, float size){
+	public DownArrow(int x, int y, float size){
 		super(x, y, size);
-				
 	}	
 	
-	protected void initPolygon(float x, float y){
+	protected void initPolygon(int x, int y){
 		polygon.reset();
 		
-		polygon.addPoint((int)x, (int)y);
+		polygon.addPoint(x, y);
 		polygon.addPoint((int)(x+size), (int)y);
 		polygon.addPoint((int)(x+size/2), (int)(y+size));
 	}
